// src/types/paths.ts
enum Paths {
  Main = "/",
  Teams = "/teams",
  TeamDetail = "/teams/", // + id
  Celebrities = "/celebrities",
  Systems = "/systems",
  Contact = "/contact",
  Search = "/search",
  Login = "/login",
  ContentCreate = "/content/create",
  Event = "/event",
  Party = "/party",
  Study = "study",
}

export default Paths;